+++ src/views/DestinationDetailView.vue
@@ -18,6 +18,7 @@ import AudioGuidePlayer from '../components/AudioGuidePlayer.vue';
 import TravelExpensesExporter from '../components/TravelExpensesExporter.vue';
 import DestinationComparator from '../components/DestinationComparator.vue';
 import CurrencyCalculator from '../components/CurrencyCalculator.vue';
+import TravelPhrasebook from '../components/TravelPhrasebook.vue';
 
 const route = useRoute();
 const travelStore = useTravelStore();
@@ -67,6 +68,7 @@ const relatedPosts = computed(() => {
         <DestinationComparator class="margin-top-calc" />
         <CurrencyCalculator class="margin-top-calc" />
         <TravelExpensesExporter :destination-id="destination.id" class="margin-top-calc" />
+        <TravelPhrasebook class="margin-top-calc" />
         <PackingChecklist class="margin-top-calc" />
       </div>